chore: CodeGenerator base class + extends
