+++ src/code-generator/PuppeteerCodeGenerator.js
@@ -1,6 +1,7 @@
 import domEvents from './dom-events-to-record'
 import pptrActions from './pptr-actions'
 import Block from './Block'
+import CodeGenerator from './CodeGenerator'
 
 const importPuppeteer = `const puppeteer = require('puppeteer');\n`
 
@@ -16,26 +17,9 @@ const wrappedHeader = `(async () => {
 const wrappedFooter = `  await browser.close()
 })()`
 
-export const defaults = {
-  wrapAsync: true,
-  headless: true,
-  waitForNavigation: true,
-  waitForSelectorOnClick: true,
-  blankLinesBetweenBlocks: true,
-  dataAttribute: '',
-  showPlaywrightFirst: false
-}
-
-export default class PuppeteerCodeGenerator {
+export default class PuppeteerCodeGenerator extends CodeGenerator {
   constructor (options) {
-    this._options = Object.assign(defaults, options)
-    this._blocks = []
-    this._frame = 'page'
-    this._frameId = 0
-    this._allFrames = {}
-    this._screenshotCounter = 1
-
-    this._hasNavigation = false
+    super(options)
   }
 
   generate (events) {
@@ -117,28 +101,6 @@ export default class PuppeteerCodeGenerator {
     return result
   }
 
-  _setFrames (frameId, frameUrl) {
-    if (frameId && frameId !== 0) {
-      this._frameId = frameId
-      this._frame = `frame_${frameId}`
-      this._allFrames[frameId] = frameUrl
-    } else {
-      this._frameId = 0
-      this._frame = 'page'
-    }
-  }
-
-  _postProcess () {
-    // when events are recorded from different frames, we want to add a frame setter near the code that uses that frame
-    if (Object.keys(this._allFrames).length > 0) {
-      this._postProcessSetFrames()
-    }
-
-    if (this._options.blankLinesBetweenBlocks && this._blocks.length > 0) {
-      this._postProcessAddBlankLines()
-    }
-  }
-
   _handleKeyDown (selector, value) {
     const block = new Block(this._frameId)
     block.addLine({ type: domEvents.KEYDOWN, value: `await ${this._frame}.type('${selector}', '${value}')` })
@@ -153,9 +115,11 @@ export default class PuppeteerCodeGenerator {
     block.addLine({ type: domEvents.CLICK, value: `await ${this._frame}.click('${selector}')` })
     return block
   }
+
   _handleChange (selector, value) {
     return new Block(this._frameId, { type: domEvents.CHANGE, value: `await ${this._frame}.select('${selector}', '${value}')` })
   }
+
   _handleGoto (href) {
     return new Block(this._frameId, { type: pptrActions.GOTO, value: `await ${this._frame}.goto('${href}')` })
   }
@@ -206,16 +170,6 @@ export default class PuppeteerCodeGenerator {
           break
         }
       }
-    }
-  }
-
-  _postProcessAddBlankLines () {
-    let i = 0
-    while (i <= this._blocks.length) {
-      const blankLine = new Block()
-      blankLine.addLine({ type: null, value: '' })
-      this._blocks.splice(i, 0, blankLine)
-      i += 2
     }
   }
 }
