+++ src/options/components/App.vue
@@ -91,7 +91,7 @@
 </template>
 
 <script>
-  import { defaults as code } from '../../code-generator/PuppeteerCodeGenerator'
+  import { defaults as code } from '../../code-generator/CodeGenerator'
 
   const defaults = {
     code,
